+++ Assets/Scripts/Entities/EntitySpawner.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 using Unity.Jobs;
 using Unity.Collections;
-using Unity.Burst;
 
 public class EntitySpawner : MonoBehaviour {
     public GameObject entityPrefab;
@@ -75,46 +74,5 @@ public class EntitySpawner : MonoBehaviour {
     public void FreePosition(Vector3 position) {
         spawnedPositions.Remove(position);
         availablePositions.Enqueue(position);
-    }
-
-    [BurstCompile]
-    struct SpawnJob : IJob {
-        [ReadOnly]
-        public NativeArray<Vector3> positions;
-        public int entitiesPerRow;
-        public float spacing;
-        public Vector3 spawnCenter;
-        public NativeArray<Vector3> result;
-
-        public void Execute() {
-            for (int i = 0; ; i++) {
-                int row = i / entitiesPerRow;
-                int col = i % entitiesPerRow;
-                Vector3 spawnPosition = CalculateSpawnPosition(col, row);
-
-                bool isValid = true;
-                for (int j = 0; j < positions.Length; j++) {
-                    if (Vector3.Distance(positions[j], spawnPosition) < spacing) {
-                        isValid = false;
-                        break;
-                    }
-                }
-
-                if (isValid) {
-                    result[0] = spawnPosition;
-                    break;
-                }
-            }
-        }
-
-        Vector3 CalculateSpawnPosition(int col, int row) {
-            Vector3 offset = new Vector3(
-                col * spacing - (entitiesPerRow - 1) * spacing / 2,
-                0, 
-                row * spacing - (entitiesPerRow - 1) * spacing / 2
-            );
-
-            return new Vector3(spawnCenter.x + offset.x, spawnCenter.y, spawnCenter.z + offset.z);
-        }
     }
 }
