+++ Assets/Scripts/Jobs and Burst/MoveJob.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e564c2f90ba3a941a8eb3402910d2b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
