+++ Assets/Scripts/Jobs and Burst/SpawnJob.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 301486c61a1404b42b1e8cc0b0d1c70d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
